Refactor imports for @hedvig-ui/Spinner
