+++ shared/hedvig-ui/index.ts
@@ -46,6 +46,7 @@ import { FlagOrbIndicator, OrbIndicator } from './OrbIndicator/orb-indicator'
 import { Popover } from './Popover/popover'
 import { RadioGroup } from './Radio/radio'
 import { SearchableDropdown } from './SearchableDropdown/searchable-dropdown'
+import { Spinner } from './Spinner/spinner'
 
 export {
   Badge,
@@ -90,4 +91,5 @@ export {
   FlagOrbIndicator,
   RadioGroup,
   SearchableDropdown,
+  Spinner,
 }
